[LWM] :lipstick: header utils components w4.0
