+++ apps/ledger-live-mobile/src/mvvm/features/Market/components/MarketListHeader/index.tsx
@@ -1,46 +1,24 @@
-import React, { useCallback } from "react";
-import { useTranslation } from "~/context/Locale";
-import { useNavigation } from "@react-navigation/native";
-import { IconButton, Text } from "@ledgerhq/lumen-ui-rnative";
-import { ArrowLeft } from "@ledgerhq/lumen-ui-rnative/symbols";
+import React from "react";
 import { useWalletFeaturesConfig } from "@ledgerhq/live-common/featureFlags/index";
+import HeaderTitle from "LLM/components/Navigation/HeaderTitle";
+import HeaderBackButton from "LLM/components/Navigation/HeaderBackButton";
 
 export function MarketListHeaderLeft() {
-  const { t } = useTranslation();
-  const navigation = useNavigation();
   const { shouldDisplayMarketBanner } = useWalletFeaturesConfig("mobile");
-
-  const handlePress = useCallback(() => {
-    navigation.goBack();
-  }, [navigation]);
 
   if (!shouldDisplayMarketBanner) {
     return null;
   }
 
-  return (
-    <IconButton
-      appearance="no-background"
-      size="md"
-      icon={ArrowLeft}
-      testID="market-list-header-left"
-      accessibilityLabel={t("common.back")}
-      onPress={handlePress}
-    />
-  );
+  return <HeaderBackButton testID="market-list-header-left" />;
 }
 
 export function MarketListHeaderTitle() {
-  const { t } = useTranslation();
   const { shouldDisplayMarketBanner } = useWalletFeaturesConfig("mobile");
 
   if (!shouldDisplayMarketBanner) {
     return null;
   }
 
-  return (
-    <Text typography="heading3SemiBold" lx={{ color: "base" }}>
-      {t("market.title")}
-    </Text>
-  );
+  return <HeaderTitle titleKey="market.title" />;
 }
